Cleaned up brought over code
Smashed mesh generation code into one single file, Meshes.jl, and brought over the Meshes test file from decapodes.
