+++ Project.toml
@@ -22,9 +22,11 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
 ACSets = "0.2"
+BenchmarkTools = "1.4.0"
 Catlab = "0.16"
 FileIO = "^1"
 GeometryBasics = "0.3, 0.4"
+JSON = "0.21.4"
 LazyArrays = "0.20, 0.21, 0.22, 1.0"
 LinearAlgebra = "1.9"
 MeshIO = "0.4"
