+++ src/CombinatorialSpaces.jl
@@ -13,17 +13,12 @@ include("FastDEC.jl")
 
 # TODO: Would be best to only have these files be included if the user wants the meshes
 include("Meshes.jl")
-include("GridMeshes.jl")
-include("SphericalMeshes.jl")
 
 @reexport using .SimplicialSets
 @reexport using .DiscreteExteriorCalculus
 @reexport using .FastDEC
 @reexport using .MeshInterop
-
 @reexport using .Meshes
-@reexport using .GridMeshes
-@reexport using .SphericalMeshes
 
 function __init__()
   @require Makie="ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a" include("MeshGraphics.jl")
